Исправляет наименование асинхронных экшенов
