+++ src/containers/favorites-container/favorites-container.jsx
@@ -3,7 +3,10 @@ import { connect } from "react-redux";
 
 import Favorites from "../../components/favorites/favorites.jsx";
 
-import { getFavorite, setFavoriteAsync } from "../../redux/fetch/actions.js";
+import {
+  getFavoriteAsync,
+  setFavoriteAsync
+} from "../../redux/hotels/actions.js";
 import { setCity } from "../../redux/hotels/actions.js";
 import { getFavoritesOffers } from "./selectors";
 
@@ -20,7 +23,7 @@ const mapStateToProps = ({ auth, hotels }) => {
 
 const mapDispatchToProps = {
   setCity,
-  getFavorite,
+  getFavoriteAsync,
   setFavoriteAsync
 };
