+++ src/containers/main-container/main-container.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { connect } from "react-redux";
 
 import Main from "../../components/main/main.jsx";
-import { setFavoriteAsync } from "../../redux/fetch/actions";
+import { setFavoriteAsync } from "../../redux/hotels/actions";
 import { setCity } from "../../redux/hotels/actions";
 import { getCityOffers, getCityCoords } from "./selectors.js";
